[Refactoring] 페이지 메타데이터 싱글톤으로 수정
