+++ pages/post/question/index.vue
